new data

+++ main/UI/UI-var.h
@@ -10,9 +10,9 @@ struct DataWgtXML {
     QString rudnik;
     QString nameVirab;
     QString izmVip;
-    int XX;
-    int YY;
-    int ZZ;
+    double XX;
+    double YY;
+    double ZZ;
 };
 
 struct DataCommission {
